// src/api/models/bible_memory_db_internal_domain_Feedback.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type bible_memory_db_internal_domain_Feedback = {
    createdAt?: string;
    id?: string;
    telegramId?: string;
    text?: string;
    updatedAt?: string;
};